Fix: Enable TPU support in AOT cache and use bfloat16 in AOT compile example
Updated get_hardware_info in helion/autotuner/aot_cache.py to detect TPUs via JAX, preventing it from incorrectly raising a "No supported GPU device found" error on TPU machines.

Switched the default dtype in examples/aot_compile_example.py from torch.float16 to torch.bfloat16. The TPU Vector ALU natively supports bfloat16, whereas forcing float16 resulted in an invalid vector type load error during Mosaic compilation.

+++ examples/aot_compile_example.py
@@ -64,8 +64,8 @@ def benchmark_add_2d() -> None:
     print(f"{'Shape':>16} {'Time (ms)':>12} {'GB/s':>10}")
     print("-" * 40)
     for m, n in shapes:
-        x = torch.randn(m, n, device=DEVICE, dtype=torch.float16)
-        y = torch.randn(m, n, device=DEVICE, dtype=torch.float16)
+        x = torch.randn(m, n, device=DEVICE, dtype=torch.bfloat16)
+        y = torch.randn(m, n, device=DEVICE, dtype=torch.bfloat16)
         add_2d(x, y)  # warmup
         time_ms = do_bench(lambda x=x, y=y: add_2d(x, y))
         assert isinstance(time_ms, float)
